Initial work on data as a trait
