+++ src/Concerns/DataTrait.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Spatie\LaravelData\Concerns;
+
+trait DataTrait
+{
+    use ResponsableData;
+    use IncludeableData;
+    use AppendableData;
+    use ValidateableData;
+    use WrapableData;
+    use TransformableData;
+    use BaseData;
+}
