+++ src/Data.php
@@ -11,7 +11,9 @@ use Illuminate\Pagination\AbstractCursorPaginator;
 use Illuminate\Pagination\AbstractPaginator;
 use Illuminate\Support\Enumerable;
 use JsonSerializable;
+use Nette\Schema\Elements\Base;
 use Spatie\LaravelData\Concerns\AppendableData;
+use Spatie\LaravelData\Concerns\DataTrait;
 use Spatie\LaravelData\Concerns\IncludeableData;
 use Spatie\LaravelData\Concerns\ResponsableData;
 use Spatie\LaravelData\Concerns\TransformableData;
@@ -32,7 +34,7 @@ use Spatie\LaravelData\Support\EloquentCasts\DataEloquentCast;
 use Spatie\LaravelData\Support\Wrapping\WrapExecutionType;
 use Spatie\LaravelData\Transformers\DataTransformer;
 
-abstract class Data implements Arrayable, Responsable, Jsonable, EloquentCastable, JsonSerializable
+abstract class Data implements DataObject
 {
     use ResponsableData;
     use IncludeableData;
@@ -40,76 +42,5 @@ abstract class Data implements Arrayable, Responsable, Jsonable, EloquentCastabl
     use ValidateableData;
     use WrapableData;
     use TransformableData;
-
-    public static function optional(mixed ...$payloads): ?static
-    {
-        if (count($payloads) === 0) {
-            return null;
-        }
-
-        foreach ($payloads as $payload) {
-            if ($payload !== null) {
-                return static::from(...$payloads);
-            }
-        }
-
-        return null;
-    }
-
-    public static function from(mixed ...$payloads): static
-    {
-        return app(DataFromSomethingResolver::class)->execute(
-            static::class,
-            ...$payloads
-        );
-    }
-
-    public static function normalizers(): array
-    {
-        return [
-            ModelNormalizer::class,
-            ArraybleNormalizer::class,
-            ObjectNormalizer::class,
-            ArrayNormalizer::class,
-        ];
-    }
-
-    public static function pipeline(): DataPipeline
-    {
-        return DataPipeline::create()
-            ->into(static::class)
-            ->through(AuthorizedDataPipe::class)
-            ->through(MapPropertiesDataPipe::class)
-            ->through(ValidatePropertiesDataPipe::class)
-            ->through(DefaultValuesDataPipe::class)
-            ->through(CastPropertiesDataPipe::class);
-    }
-
-    public static function collection(Enumerable|array|AbstractPaginator|AbstractCursorPaginator|Paginator|DataCollection $items): DataCollection|PaginatedDataCollection
-    {
-        $isPaginated = $items instanceof AbstractPaginator
-            || $items instanceof AbstractCursorPaginator
-            || $items instanceof Paginator;
-
-        return $isPaginated
-            ? new PaginatedDataCollection(static::class, $items)
-            : new DataCollection(static::class, $items);
-    }
-
-    public static function empty(array $extra = []): array
-    {
-        return app(EmptyDataResolver::class)->execute(static::class, $extra);
-    }
-
-    public function transform(
-        bool $transformValues = true,
-        WrapExecutionType $wrapExecutionType = WrapExecutionType::Disabled,
-    ): array {
-        return DataTransformer::create($transformValues, $wrapExecutionType)->transform($this);
-    }
-
-    public static function castUsing(array $arguments)
-    {
-        return new DataEloquentCast(static::class);
-    }
+    use DataTrait;
 }
